Adds isSubsequence solution
Implements a two-pointer solution to determine if one string is a subsequence of another.

The solution iterates through the longer string and compares each character with the current character of the potential subsequence.
If a match is found, the subsequence index is incremented.
The function returns true if the subsequence index reaches the end of the subsequence string, indicating that it is indeed a subsequence.
Returns false is the given string t is empty or when the subsequence length is longer than the given string t.