Adds solution for kids with candies problem
Implements a function to determine which kids can have the greatest number of candies after receiving extra candies.

The solution iterates through the candies array and checks if adding the extra candies to each kid's current candy count results in a value greater than or equal to the maximum candy count in the array.